aider: Reversed the order of components in `generateComponents.js` and `server.js` to ensure the most recently added component appears on top.
# Aider chat conversation:

ASSISTANT: To make the most recently added component appear on top, we need to modify the `generateComponents.js` and `server.js` files.

In `generateComponents.js`, we need to reverse the order of the components after they are sorted by time.

In `server.js`, we need to reverse the order of the components before they are sent to the client.

Here are the *SEARCH/REPLACE* blocks:

generateComponents.js
```javascript
<<<<<<< SEARCH
    .map(dirent => dirent.name);
=======
    .map(dirent => dirent.name)
    .reverse();
>>>>>>> REPLACE
```

server.js
```javascript
<<<<<<< SEARCH
  res.send(componentDirs);
=======
  res.send(componentDirs.reverse());
>>>>>>> REPLACE
```

These changes will ensure that the most recently added component is always at the top of the list.