aider: Fix fetch request syntax error and ensure then() method is called when if condition is true.
# Aider chat conversation:

ASSISTANT: Looking at the code around line 126, it seems like the issue might be related to the fetch request inside the if condition. The fetch request is commented out, but the following then() method calls are not, which could be causing a syntax error.

Here's how to fix it:

public/app.js
```javascript
<<<<<<< SEARCH
  if (componentName && htmlCode && cssCode && jsCode) {
                        fetch("/save-component", {
                                method: "POST",
                                headers: {
                                        "Content-Type": "application/json",
                                },
                                body: JSON.stringify(componentData),
                        })
                } else {
                        alert('All fields are required');
                }

		/* fetch("/save-component", {
			method: "POST",
			headers: {
				"Content-Type": "application/json",
			},
			body: JSON.stringify(componentData),
		}) */

			.then((response) => response.json())
=======
  if (componentName && htmlCode && cssCode && jsCode) {
    fetch("/save-component", {
      method: "POST",
      headers: {
        "Content-Type": "application/json",
      },
      body: JSON.stringify(componentData),
    })
    .then((response) => response.json())
  } else {
    alert('All fields are required');
  }
>>>>>>> REPLACE
```

This change will ensure that the then() method is called on the fetch request when the if condition is true. If the condition is false, an alert will be shown, and no fetch request will be made.