aider: Fix screenshot generation route to replace hyphens with spaces in the component name before accessing the file.
# Aider chat conversation:

USER: when saving new component i get screenshot with error message: Cannot GET /components/opet-dvije/opet-dvije.html. probabaly because of adding "-" after screenshot is make, or?

ASSISTANT: Yes, it seems like the issue is due to the hyphen ("-") being added to the component name when generating the screenshot. The screenshot generation is trying to access the component's HTML file using the hyphenated name, but the file is saved with spaces in the name, not hyphens.

To fix this, we need to modify `server.js`:

1. Update the screenshot generation route to replace spaces with hyphens in the component name before accessing the file.

Here is the *SEARCH/REPLACE* block:

server.js
```python
<<<<<<< SEARCH
app.post('/generate-screenshot/:name', async (req, res) => {
	const componentName = req.params.name;
=======
app.post('/generate-screenshot/:name', async (req, res) => {
	const componentName = req.params.name.replace(/-/g, " ");
>>>>>>> REPLACE
```

This change will ensure that when the screenshot generation route receives a component name with hyphens, it will replace those hyphens with spaces before trying to access the component's HTML file.