aider: Add else condition to response.json() in app.js
# Aider chat conversation:

USER: error: Uncaught SyntaxError: Unexpected token '.' (at app.js:165:4)

ASSISTANT: The error message indicates that there is a syntax error in your JavaScript code at line 165 in the `app.js` file. The error is caused by an unexpected token '.', which usually means that there is a misplaced or unexpected character in your code.

Looking at the code you provided, it seems like the error is caused by the following line:

```javascript
.then((response) => response.json()) : // Add else condition
```

This line is incomplete and is causing the syntax error. It seems like you are trying to use a ternary operator, but the else part of the operator is missing.

To fix this issue, you need to complete the ternary operator with the appropriate else condition. If you are not sure what to put in the else condition, you can use an empty function as a placeholder.

Here is the *SEARCH/REPLACE* block:

public/app.js
```javascript
<<<<<<< SEARCH
.then((response) => response.json()) : // Add else condition
=======
.then((response) => response.json()) : () => {} // Add else condition
>>>>>>> REPLACE
```

Please replace the above code block in your `app.js` file. This should resolve the syntax error.