aider: Fix WebSocket connection not pushing updates to the client when a new component is added.
# Aider chat conversation:

ASSISTANT: It seems like the issue is related to the WebSocket connection not pushing updates to the client when a new component is added.

To fix this, we need to modify the `public/app.js` and `server.js` files.

In `public/app.js`, we need to add a WebSocket message handler to reload the page when a new component is added.

In `server.js`, we need to send a WebSocket message to all connected clients when a new component is added.

Here are the *SEARCH/REPLACE* blocks:

public/app.js
```javascript
<<<<<<< SEARCH
socket.onmessage = function (event) {
	console.log(`[message] Data received from server: ${event.data}`);
};
=======
socket.onmessage = function (event) {
	console.log(`[message] Data received from server: ${event.data}`);
	if (event.data === 'componentAdded') {
		window.location.reload();
	}
};
>>>>>>> REPLACE
```

server.js
```javascript
<<<<<<< SEARCH
fs.watch('./public/components', (eventType, filename) => {
  const componentHTML = updateComponents();
  wss.broadcast(JSON.stringify(componentHTML));
});
=======
fs.watch('./public/components', (eventType, filename) => {
  const componentHTML = updateComponents();
  wss.broadcast('componentAdded');
});
>>>>>>> REPLACE
```

These changes will cause the page to reload whenever a new component is added, ensuring that the new component is displayed immediately.