aider: Fix issue with adding tags by modifying the route handler for '/tags/' in server.js.
# Aider chat conversation:

USER: When attempting to add a tag, the following error arises: "POST http://localhost:3000/tags/ 404 (Not Found)". This issue is located at line 397 in app.js

ASSISTANT: The error message indicates that the server is not able to find the route for the POST request to '/tags/'. This is likely because the server-side code (in server.js) does not have a route handler for '/tags/'.

To fix this, we need to modify the `server.js` file to add a route handler for '/tags/'.

Here is the *SEARCH/REPLACE* block:

server.js
```javascript
<<<<<<< SEARCH
app.post('/tags/:componentId', (req, res) => {
=======
app.post('/tags/', (req, res) => {
>>>>>>> REPLACE
```

This change will make the server respond to POST requests at '/tags/' instead of '/tags/:componentId'. If you still need to handle requests at '/tags/:componentId', you should add a new route handler for '/tags/' instead of modifying the existing one.